+++ src/test/ui/utils/actions/custom-actions/triggerPageFunctionalTests.action.ts
@@ -8,7 +8,7 @@ import {
   enable_error_message_validation,
   enable_navigation_tests,
   enable_visibility_validation,
-} from '../../../../../../playwright.sauce.config';
+} from '../../../../../../playwright.config';
 import { shortUrl, truncateForLog } from '../../common/string.utils';
 import { IAction } from '../../interfaces';
 import {
