+++ src/test/ui/utils/common/log-test-env.ts
@@ -1,4 +1,4 @@
-import { enable_pft_debug_log } from '../../../../../playwright.sauce.config';
+import { enable_pft_debug_log } from '../../../../../playwright.config';
 
 /** Env keys set in `respondToAClaimCrossBrowser.spec.ts` `beforeEach` (plus `CASE_NUMBER` from submit). */
 export const RESPOND_TO_CLAIM_BEFORE_EACH_ENV_KEYS = [
